+++ gammapy/cube/__init__.py
@@ -1,7 +1,6 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 """Sky cubes (3-dimensional: energy, lon, lat)."""
 from .background import *
-from .edisp_map import *
 from .exposure import *
 from .fov import *
 from .make import *
