+++ gammapy/datasets/map.py
@@ -7,7 +7,7 @@ from astropy.nddata.utils import NoOverlapError
 from astropy.table import Table
 from astropy.utils import lazyproperty
 from regions import CircleSkyRegion
-from gammapy.cube.edisp_map import EDispMap
+from gammapy.irf.edisp_map import EDispMap
 from gammapy.cube.psf_kernel import PSFKernel
 from gammapy.cube.psf_map import PSFMap
 from gammapy.data import GTI
