+++ gammapy/irf/edisp_map.py
@@ -7,53 +7,7 @@ from gammapy.irf import EDispKernel
 from gammapy.maps import Map, MapAxis, MapCoord, WcsGeom
 from gammapy.utils.random import InverseCDFSampler, get_random_state
 
-__all__ = ["make_edisp_map", "EDispMap"]
-
-
-def make_edisp_map(edisp, pointing, geom, exposure_map=None):
-    """Make a edisp map for a single observation
-
-    Expected axes : migra and true energy in this specific order
-    The name of the migra MapAxis is expected to be 'migra'
-
-    Parameters
-    ----------
-    edisp : `~gammapy.irf.EnergyDispersion2D`
-        the 2D Energy Dispersion IRF
-    pointing : `~astropy.coordinates.SkyCoord`
-        the pointing direction
-    geom : `~gammapy.maps.Geom`
-        the map geom to be used. It provides the target geometry.
-        rad and true energy axes should be given in this specific order.
-    exposure_map : `~gammapy.maps.Map`, optional
-        the associated exposure map.
-        default is None
-
-    Returns
-    -------
-    edispmap : `~gammapy.cube.EDispMap`
-        the resulting EDisp map
-    """
-    energy_axis = geom.get_axis_by_name("energy_true")
-    energy = energy_axis.center
-
-    migra_axis = geom.get_axis_by_name("migra")
-    migra = migra_axis.center
-
-    # Compute separations with pointing position
-    offset = geom.separation(pointing)
-
-    # Compute EDisp values
-    edisp_values = edisp.data.evaluate(
-        offset=offset,
-        energy_true=energy[:, np.newaxis, np.newaxis, np.newaxis],
-        migra=migra[:, np.newaxis, np.newaxis],
-    )
-
-    # Create Map and fill relevant entries
-    data = edisp_values.to_value("")
-    edispmap = Map.from_geom(geom, data=data, unit="")
-    return EDispMap(edispmap, exposure_map)
+__all__ = ["EDispMap"]
 
 
 class EDispMap:
